+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 
 ---
+## [5.1.0] - 2022-08-29
+### Added
+- `write_array::<Length>()`
+- `read_array::<Length>()`
+### Changed
+- `read_array()` -> `read_vec()`
+- `write_array()` -> `write_vec()`
+### Fixed
+### Removed
+---
+
+---
 ## [5.0.0] - 2022-08-24
 ### Added
 ### Changed